Factor out search suggestions into a hook
