+++ src/pages/search/index.tsx
@@ -4,6 +4,7 @@ import { useQuery } from "@apollo/react-hooks";
 import { Helmet } from "react-helmet";
 
 import { useScrollTopOnMount } from "src/shared/hooks/useScrollTopOnMount";
+import { useSearchSuggestions } from "src/shared/hooks/useSearchSuggestions";
 import { useSearchParams } from "src/shared/hooks/useSearchParams";
 import {
   useSearch,
@@ -13,19 +14,14 @@ import {
 import { SearchType, RESULTS_PER_PAGE } from "src/shared/constants/search";
 import pageCopy from "./copy";
 
-import { GetSearchSuggestions } from "./graphql/types/GetSearchSuggestions";
 import { GetAllSearch } from "./graphql/types/GetAllSearch";
 import {
-  GET_SEARCH_SUGGESTIONS,
   GET_ALL_SEARCH,
   GET_COMPANIES_SEARCH,
   GET_JOBS_SEARCH,
   GET_REVIEWS_SEARCH,
 } from "./graphql/queries";
-import {
-  buildSearchSuggestions,
-  buildSearchResultCardsList,
-} from "./graphql/utils";
+import { buildSearchResultCardsList } from "./graphql/utils";
 
 import {
   ResultCardDisplay,
@@ -134,19 +130,7 @@ export const Heading = styled(Text)`
 const GenericSearchPage: React.FC = () => {
   useScrollTopOnMount();
 
-  /**
-   * Fetch the data we need for suggestions
-   */
-  const {
-    // loading: suggestionsLoading,
-    // error: suggestionsError,
-    data: suggestionsData,
-  } = useQuery<GetSearchSuggestions>(GET_SEARCH_SUGGESTIONS);
-
-  const allSuggestions = useMemo(
-    () => buildSearchSuggestions(suggestionsData),
-    [suggestionsData]
-  );
+  const searchSuggestions = useSearchSuggestions();
 
   const {
     // for fetching results
@@ -212,7 +196,7 @@ const GenericSearchPage: React.FC = () => {
       <PageContainer id="search-page">
         <Heading variant="heading1">{headingMarkup}</Heading>
         <SearchField
-          suggestions={allSuggestions}
+          suggestions={searchSuggestions}
           onTriggerSearch={onNewSearch}
         />
         <ResultCardDisplay
