+++ src/shared/hooks/useSearchSuggestions/graphql/queries.ts
@@ -1,0 +1,17 @@
+import { gql } from "apollo-boost";
+
+export const GET_SEARCH_SUGGESTIONS = gql`
+  query GetSearchSuggestions {
+    companiesList {
+      items {
+        name
+      }
+    }
+
+    jobsList {
+      items {
+        name
+      }
+    }
+  }
+`;
