+++ src/shared/hooks/useSearchSuggestions/graphql/types/GetSearchSuggestions.ts
